- MeasuresChecking: now it is possible to compute one single measure for all the constraints and traces, without initializing the entire megaMatrix monster - Variant analysis corrected p_value calculation - Variant analysis now uses the Float matrixMonster. - general refactoring

+++ src/minerful/JanusVariantAnalysisStarter.java
@@ -108,7 +108,7 @@ public class JanusVariantAnalysisStarter extends MinerFulMinerStarter {
         }
 
         JanusVariantAnalysisLauncher variantAnalysis = new JanusVariantAnalysisLauncher(janusParams, systemParams);
-        Map<String, Double> result = variantAnalysis.checkVariants();
+        Map<String, Float> result = variantAnalysis.checkVariants();
 
         TaskCharArchive alphabet = variantAnalysis.getAlphabetDecoder();
         new JanusVariantOutputManagementLauncher().manageVariantOutput(result, viewParams, janusParams, systemParams, alphabet);
